Refactor into lib
Initial step in the process of making doug more maintainable and idiomatic.
